+++ programs/drift/src/math/margin/tests.rs
@@ -2494,6 +2494,9 @@ mod validate_spot_margin_trading {
     use crate::test_utils::get_pyth_price;
     use crate::test_utils::*;
 
+    use crate::state::perp_market::{ContractTier, MarketStatus, PerpMarket, AMM};
+    use crate::{AMM_RESERVE_PRECISION, BASE_PRECISION_I64, PEG_PRECISION, QUOTE_PRECISION_I64};
+
     #[test]
     pub fn sol_ask_larger_than_deposit() {
         let slot = 0_u64;
@@ -2510,7 +2513,7 @@ mod validate_spot_margin_trading {
         );
         let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
 
-        let _market_map = PerpMarketMap::empty();
+        let perp_market_map = PerpMarketMap::empty();
 
         let mut usdc_spot_market = SpotMarket {
             market_index: 0,
@@ -2567,7 +2570,12 @@ mod validate_spot_margin_trading {
             ..User::default()
         };
 
-        let result = validate_spot_margin_trading(&user, &spot_market_map, &mut oracle_map);
+        let result = validate_spot_margin_trading(
+            &user,
+            &perp_market_map,
+            &spot_market_map,
+            &mut oracle_map,
+        );
 
         assert_eq!(result, Err(ErrorCode::MarginTradingDisabled));
     }
@@ -2588,7 +2596,7 @@ mod validate_spot_margin_trading {
         );
         let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
 
-        let _market_map = PerpMarketMap::empty();
+        let perp_market_map = PerpMarketMap::empty();
 
         let mut usdc_spot_market = SpotMarket {
             market_index: 0,
@@ -2645,7 +2653,12 @@ mod validate_spot_margin_trading {
             ..User::default()
         };
 
-        let result = validate_spot_margin_trading(&user, &spot_market_map, &mut oracle_map);
+        let result = validate_spot_margin_trading(
+            &user,
+            &perp_market_map,
+            &spot_market_map,
+            &mut oracle_map,
+        );
 
         assert_eq!(result, Ok(()));
     }
@@ -2666,7 +2679,7 @@ mod validate_spot_margin_trading {
         );
         let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
 
-        let _market_map = PerpMarketMap::empty();
+        let perp_market_map = PerpMarketMap::empty();
 
         let mut usdc_spot_market = SpotMarket {
             market_index: 0,
@@ -2723,7 +2736,12 @@ mod validate_spot_margin_trading {
             ..User::default()
         };
 
-        let result = validate_spot_margin_trading(&user, &spot_market_map, &mut oracle_map);
+        let result = validate_spot_margin_trading(
+            &user,
+            &perp_market_map,
+            &spot_market_map,
+            &mut oracle_map,
+        );
 
         assert_eq!(result, Err(ErrorCode::MarginTradingDisabled));
     }
@@ -2744,7 +2762,7 @@ mod validate_spot_margin_trading {
         );
         let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
 
-        let _market_map = PerpMarketMap::empty();
+        let perp_market_map = PerpMarketMap::empty();
 
         let mut usdc_spot_market = SpotMarket {
             market_index: 0,
@@ -2801,7 +2819,12 @@ mod validate_spot_margin_trading {
             ..User::default()
         };
 
-        let result = validate_spot_margin_trading(&user, &spot_market_map, &mut oracle_map);
+        let result = validate_spot_margin_trading(
+            &user,
+            &perp_market_map,
+            &spot_market_map,
+            &mut oracle_map,
+        );
 
         assert_eq!(result, Err(ErrorCode::MarginTradingDisabled));
     }
@@ -2822,7 +2845,7 @@ mod validate_spot_margin_trading {
         );
         let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
 
-        let _market_map = PerpMarketMap::empty();
+        let perp_market_map = PerpMarketMap::empty();
 
         let mut usdc_spot_market = SpotMarket {
             market_index: 0,
@@ -2879,7 +2902,12 @@ mod validate_spot_margin_trading {
             ..User::default()
         };
 
-        let result = validate_spot_margin_trading(&user, &spot_market_map, &mut oracle_map);
+        let result = validate_spot_margin_trading(
+            &user,
+            &perp_market_map,
+            &spot_market_map,
+            &mut oracle_map,
+        );
 
         assert_eq!(result, Ok(()));
     }
@@ -2900,7 +2928,7 @@ mod validate_spot_margin_trading {
         );
         let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
 
-        let _market_map = PerpMarketMap::empty();
+        let perp_market_map = PerpMarketMap::empty();
 
         let mut usdc_spot_market = SpotMarket {
             market_index: 0,
@@ -2957,9 +2985,128 @@ mod validate_spot_margin_trading {
             ..User::default()
         };
 
-        let result = validate_spot_margin_trading(&user, &spot_market_map, &mut oracle_map);
+        let result = validate_spot_margin_trading(
+            &user,
+            &perp_market_map,
+            &spot_market_map,
+            &mut oracle_map,
+        );
 
         assert_eq!(result, Err(ErrorCode::MarginTradingDisabled));
+    }
+
+    #[test]
+    pub fn attempt_enable_margin_trading_with_isolated_perp() {
+        let slot = 0_u64;
+
+        let mut sol_oracle_price = get_pyth_price(100, 6);
+        let sol_oracle_price_key =
+            Pubkey::from_str("J83w4HKfqxwcq3BEMMkPFSppX3gqekLyLJBexebFVkix").unwrap();
+        let pyth_program = crate::ids::pyth_program::id();
+        create_account_info!(
+            sol_oracle_price,
+            &sol_oracle_price_key,
+            &pyth_program,
+            oracle_account_info
+        );
+        let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
+
+        let mut market = PerpMarket {
+            amm: AMM {
+                base_asset_reserve: 100 * AMM_RESERVE_PRECISION,
+                quote_asset_reserve: 100 * AMM_RESERVE_PRECISION,
+                bid_base_asset_reserve: 101 * AMM_RESERVE_PRECISION,
+                bid_quote_asset_reserve: 99 * AMM_RESERVE_PRECISION,
+                ask_base_asset_reserve: 99 * AMM_RESERVE_PRECISION,
+                ask_quote_asset_reserve: 101 * AMM_RESERVE_PRECISION,
+                sqrt_k: 100 * AMM_RESERVE_PRECISION,
+                peg_multiplier: 100 * PEG_PRECISION,
+                order_step_size: 10000000,
+                oracle: sol_oracle_price_key,
+                ..AMM::default()
+            },
+            margin_ratio_initial: 1000,
+            margin_ratio_maintenance: 500,
+            status: MarketStatus::Initialized,
+            contract_tier: ContractTier::Isolated,
+
+            ..PerpMarket::default()
+        };
+        create_anchor_account_info!(market, PerpMarket, market_account_info);
+        let perp_market_map = PerpMarketMap::load_one(&market_account_info, true).unwrap();
+
+        let mut usdc_spot_market = SpotMarket {
+            market_index: 0,
+            oracle_source: OracleSource::QuoteAsset,
+            cumulative_deposit_interest: SPOT_CUMULATIVE_INTEREST_PRECISION,
+            decimals: 6,
+            initial_asset_weight: SPOT_WEIGHT_PRECISION,
+            maintenance_asset_weight: SPOT_WEIGHT_PRECISION,
+            deposit_balance: 10000 * SPOT_BALANCE_PRECISION,
+            liquidator_fee: 0,
+            ..SpotMarket::default()
+        };
+        create_anchor_account_info!(usdc_spot_market, SpotMarket, usdc_spot_market_account_info);
+        let mut sol_spot_market = SpotMarket {
+            market_index: 1,
+            oracle_source: OracleSource::Pyth,
+            oracle: sol_oracle_price_key,
+            cumulative_deposit_interest: SPOT_CUMULATIVE_INTEREST_PRECISION,
+            cumulative_borrow_interest: SPOT_CUMULATIVE_INTEREST_PRECISION,
+            decimals: 9,
+            initial_asset_weight: 8 * SPOT_WEIGHT_PRECISION / 10,
+            maintenance_asset_weight: 9 * SPOT_WEIGHT_PRECISION / 10,
+            initial_liability_weight: 12 * SPOT_WEIGHT_PRECISION / 10,
+            maintenance_liability_weight: 11 * SPOT_WEIGHT_PRECISION / 10,
+            liquidator_fee: LIQUIDATION_FEE_PRECISION / 1000,
+            ..SpotMarket::default()
+        };
+        create_anchor_account_info!(sol_spot_market, SpotMarket, sol_spot_market_account_info);
+        let spot_market_account_infos = Vec::from([
+            &usdc_spot_market_account_info,
+            &sol_spot_market_account_info,
+        ]);
+        let spot_market_map =
+            SpotMarketMap::load_multiple(spot_market_account_infos, true).unwrap();
+
+        let mut spot_positions = [SpotPosition::default(); 8];
+        spot_positions[0] = SpotPosition {
+            market_index: 0,
+            balance_type: SpotBalanceType::Deposit,
+            scaled_balance: 100 * SPOT_BALANCE_PRECISION_U64,
+            ..SpotPosition::default()
+        };
+        spot_positions[1] = SpotPosition {
+            market_index: 1,
+            balance_type: SpotBalanceType::Deposit,
+            scaled_balance: 2 * SPOT_BALANCE_PRECISION_U64,
+            open_bids: 2 * 10_i64.pow(9),
+            ..SpotPosition::default()
+        };
+        let user = User {
+            orders: [Order::default(); 32],
+            perp_positions: get_positions(PerpPosition {
+                market_index: 0,
+                base_asset_amount: BASE_PRECISION_I64,
+                quote_asset_amount: -150 * QUOTE_PRECISION_I64,
+                quote_entry_amount: -150 * QUOTE_PRECISION_I64,
+                quote_break_even_amount: -150 * QUOTE_PRECISION_I64,
+                open_orders: 1,
+                open_bids: BASE_PRECISION_I64,
+                ..PerpPosition::default()
+            }),
+            spot_positions,
+            ..User::default()
+        };
+
+        let result = validate_spot_margin_trading(
+            &user,
+            &perp_market_map,
+            &spot_market_map,
+            &mut oracle_map,
+        );
+
+        assert_eq!(result, Err(ErrorCode::IsolatedAssetTierViolation));
     }
 }