+++ programs/drift/src/state/oracle_map.rs
@@ -1,6 +1,8 @@
 use crate::error::ErrorCode::UnableToLoadOracle;
 use crate::error::{DriftResult, ErrorCode};
-use crate::ids::{bonk_oracle, pepe_oracle, pyth_program, usdc_oracle, usdt_oracle_mainnet};
+use crate::ids::{
+    bonk_oracle, pepe_oracle, pyth_program, switchboard_program, usdc_oracle, usdt_oracle_mainnet,
+};
 use crate::math::constants::PRICE_PRECISION_I64;
 use crate::math::oracle::{oracle_validity, OracleValidity};
 use crate::state::oracle::{get_oracle_price, OraclePriceData, OracleSource, PrelaunchOracle};
@@ -85,6 +87,7 @@ impl<'a> OracleMap<'a> {
         market_index: u16,
         pubkey: &Pubkey,
         last_oracle_price_twap: i64,
+        max_confidence_interval_multiplier: u64,
     ) -> DriftResult<(&OraclePriceData, OracleValidity)> {
         if self.should_get_quote_asset_price_data(pubkey) {
             return Ok((&self.quote_asset_price_data, OracleValidity::Valid));
@@ -102,6 +105,7 @@ impl<'a> OracleMap<'a> {
                     last_oracle_price_twap,
                     oracle_price_data,
                     &self.oracle_guard_rails.validity,
+                    max_confidence_interval_multiplier,
                     true,
                 )?;
                 self.validity.insert(*pubkey, oracle_validity);
@@ -132,6 +136,7 @@ impl<'a> OracleMap<'a> {
             last_oracle_price_twap,
             oracle_price_data,
             &self.oracle_guard_rails.validity,
+            max_confidence_interval_multiplier,
             true,
         )?;
         self.validity.insert(*pubkey, oracle_validity);
@@ -233,6 +238,19 @@ impl<'a> OracleMap<'a> {
                 );
 
                 continue;
+            } else if account_info.owner == &switchboard_program::id() {
+                let account_info = account_info_iter.next().safe_unwrap()?;
+                let pubkey = account_info.key();
+
+                oracles.insert(
+                    pubkey,
+                    AccountInfoAndOracleSource {
+                        account_info: account_info.clone(),
+                        oracle_source: OracleSource::Switchboard,
+                    },
+                );
+
+                continue;
             }
 
             break;
@@ -301,12 +319,20 @@ impl<'a> OracleMap<'a> {
             }
 
             let pubkey = account_info.key();
-
             oracles.insert(
                 pubkey,
                 AccountInfoAndOracleSource {
                     account_info: account_info.clone(),
                     oracle_source: OracleSource::Prelaunch,
+                },
+            );
+        } else if account_info.owner == &switchboard_program::id() {
+            let pubkey = account_info.key();
+            oracles.insert(
+                pubkey,
+                AccountInfoAndOracleSource {
+                    account_info: account_info.clone(),
+                    oracle_source: OracleSource::Switchboard,
                 },
             );
         } else if account_info.key() != Pubkey::default() {